refactor(#27): ChatController 인프라 로직 분리 및 LobbyController 서비스 레이어 경유 구조 수정
- ChatService 신규 생성: extractUserIdentifier(), buildSecureMessage() 로직 이전
- LobbyService 신규 생성: LobbyController의 Repository 직접 참조 제거
- ChatController: ChatService 위임으로 순수 라우팅만 담당하도록 슬림화
- LobbyController: LobbyRepository 직접 주입 → LobbyService 경유로 수정
